+++ vent/vent.go
@@ -199,6 +199,8 @@ type K8PodEnv struct {
 	Env map[string]string `json:"env"`
 }
 
+// processItem looks up key in the indexer, converts it into a v1.Pod,
+// and calls PostToWebhooks.
 func (c *Controller) processItem(key string) error {
 	c.logger.Infof("Processing change to Pod %s", key)
 
@@ -207,13 +209,16 @@ func (c *Controller) processItem(key string) error {
 		return fmt.Errorf("Error fetching object with key %s from store: %v", key, err)
 	}
 	pod := v1.Pod{}
+	webhookURLs := c.urls
 	if exists {
-		objJSON, jsonErr := json.Marshal(obj)
-		if jsonErr != nil {
-			return fmt.Errorf("failed to marshal object to JSON:%v", jsonErr)
+		var extractErr error
+		var webhooks []string
+		pod, webhooks, extractErr = extractPod(obj, c.logger)
+		if extractErr != nil {
+			return extractErr
 		}
-		if err := json.Unmarshal(objJSON, &pod); err != nil {
-			return fmt.Errorf("failed to unmarshal object as Pod: %v", err)
+		if webhooks != nil && len(webhooks) > 0 {
+			webhookURLs = webhooks
 		}
 	} else {
 		splitName := strings.SplitN(key, "/", 2)
@@ -232,7 +237,47 @@ func (c *Controller) processItem(key string) error {
 		Pod: pod,
 		Env: c.env,
 	}
-	PostToWebhooks(c.urls, postIt)
+	PostToWebhooks(webhookURLs, postIt)
 
 	return nil
+}
+
+// K8VentPodAnnotation defines the valid structure of the
+// "atomist.com/k8vent" pod annotation.
+type K8VentPodAnnotation struct {
+	Webhooks []string `json:"webhooks"`
+}
+
+// extractPod tries to convert object into a v1.Pod by marshaling it
+// to JSON and back again, returning it as p.  If there is an
+// "atomist.com/k8vent" annotation on the pod, it parses it as JSON
+// and returns the value of the webhooks key, if it exists, as w.  If
+// an error occurs, e will be non-nil.
+func extractPod(obj interface{}, logger *logrus.Entry) (p v1.Pod, w []string, e error) {
+	objJSON, jsonErr := json.Marshal(obj)
+	if jsonErr != nil {
+		return p, nil, fmt.Errorf("failed to marshal object to JSON: %v", jsonErr)
+	}
+	pod := v1.Pod{}
+	if err := json.Unmarshal(objJSON, &pod); err != nil {
+		return p, nil, fmt.Errorf("failed to unmarshal object as Pod: %v", err)
+	}
+
+	ventAnnot, annotOK := pod.Annotations["atomist.com/k8vent"]
+	if !annotOK {
+		return pod, nil, nil
+	}
+
+	annot := K8VentPodAnnotation{}
+	if err := json.Unmarshal([]byte(ventAnnot), &annot); err != nil {
+		logger.Infof("failed to unmarshal k8vent annotation '%s': %v", ventAnnot, err)
+		return pod, nil, nil
+	}
+
+	if len(annot.Webhooks) > 0 {
+		return pod, annot.Webhooks, nil
+	}
+	logger.Infof("webhooks parsed from %s/%s but zero length", pod.Namespace, pod.Name)
+
+	return pod, nil, nil
 }