+++ CHANGELOG.md
@@ -5,7 +5,20 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/atomist/k8vent/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/atomist/k8vent/compare/0.2.0...HEAD
+
+## [0.2.0]
+
+[0.2.0]: https://github.com/atomist/k8vent/compare/0.1.0...0.2.0
+
+Multi-tenant release
+
+### Added
+
+-   You can provide pod-specific webhook URLS in the
+    "atomist.com/k8vent" pod annotation [#2][2]
+
+[2]: https://github.com/atomist/k8vent/issues/2
 
 ## [0.1.0] - 2018-01-04